refactoring tiles, offline, geoman. WIP still
